% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clone_qc}
\alias{plot_clone_qc}
\title{Plot the QC pairwise rule matrix}
\usage{
plot_clone_qc(
  qc,
  ccf_label = c("max", "mean", "none"),
  filename = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{qc}{A \code{clonetree_qc} object.}

\item{ccf_label}{Which cluster-level CCF summary to show under each cluster
name on both axes.}

\item{filename, width, height}{Optional save arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the QC pairwise rule matrix
}
